+++ Avatars/Runtime/Scripts/Controllers/AvatarControllerBase.cs
@@ -19,8 +19,11 @@ namespace Reflectis.SDK.Avatars
         [SerializeField, Tooltip("\"This\" character, i.e. the character (avatar) that is associated to this avatar controller")]
         protected CharacterBase characterReference;
 
-        [SerializeField, Tooltip("Reference to the voice chat audio source")] 
+        [SerializeField, Tooltip("Reference to the voice chat audio source")]
         protected AudioSource voiceChatSource;
+
+        [SerializeField, Tooltip("Reference to the voice chat audio source")]
+        protected Transform avatarContainer;
 
         #endregion
 
@@ -35,6 +38,7 @@ namespace Reflectis.SDK.Avatars
         public CharacterBase CharacterReference { get => characterReference; private set => characterReference = value; }
         public CharacterControllerBase SourceCharacterController { get; private set; }
         public AudioSource VoiceChatSource { get => voiceChatSource; }
+        public Transform AvatarContainer => avatarContainer;
 
         #endregion