+++ ClientModels/Runitime/Scripts/Models/CMEnvironment.cs
@@ -1,5 +1,6 @@
+using Reflectis.SDK.Platform;
 using System;
-
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace Reflectis.SDK.ClientModels
@@ -17,16 +18,35 @@ namespace Reflectis.SDK.ClientModels
         [SerializeField] private int worldId;
         [SerializeField] private string localizationName = "localizationcsv";
         [SerializeField] private TextAsset localizationCSV;
+        [SerializeField] private List<ESupportedPlatform> platforms;
 
         public int ID { get => id; set => id = value; }
         public string Name { get => name; set => name = value; }
         public string Description { get => description; set => description = value; }
-        public string ImageUrl { get => imageUrl; set => imageUrl = value; }
+        public string AddressableImageName { get => imageUrl; set => imageUrl = value; }
         public Texture ImageTexture { get => imageTexture; set => imageTexture = value; }
         public string AddressableKey { get => addressableKey; set => addressableKey = value; }
         public string Catalog { get => catalog; set => catalog = value; }
         public int WorldId { get => worldId; set => worldId = value; }
         public string LocalizationName { get => localizationName; set => localizationName = value; }
         public TextAsset LocalizationCSV { get => localizationCSV; set => localizationCSV = value; }
+        public List<ESupportedPlatform> Platforms { get => platforms; set => platforms = value; }
+
+        public string PlatformsListString
+        {
+            get
+            {
+                string platforms = "";
+                foreach (var eventDataPlatforms in Platforms)
+                {
+                    platforms += eventDataPlatforms.ToString() + " / ";
+                }
+                if (platforms.Length > 0)
+                {
+                    platforms = platforms.Substring(0, platforms.Length - " / ".Length);
+                }
+                return platforms;
+            }
+        }
     }
 }